refactored the folder and file structure for kinds and types (just moved code around, no further changes)
