+++ packages/typir/src/index.ts
@@ -21,13 +21,28 @@ export * from './initialization/type-initializer.js';
 export * from './initialization/type-reference.js';
 export * from './initialization/type-selector.js';
 export * from './initialization/type-waiting.js';
-export * from './kinds/class-kind.js';
-export * from './kinds/fixed-parameters-kind.js';
-export * from './kinds/function-kind.js';
+export * from './kinds/bottom/bottom-kind.js';
+export * from './kinds/bottom/bottom-type.js';
+export * from './kinds/class/class-initializer.js';
+export * from './kinds/class/class-kind.js';
+export * from './kinds/class/class-type.js';
+export * from './kinds/class/class-validation.js';
+export * from './kinds/class/top-class-kind.js';
+export * from './kinds/class/top-class-type.js';
+export * from './kinds/fixed-parameters/fixed-parameters-kind.js';
+export * from './kinds/fixed-parameters/fixed-parameters-type.js';
+export * from './kinds/function/function-initializer.js';
+export * from './kinds/function/function-kind.js';
+export * from './kinds/function/function-type.js';
+export * from './kinds/function/function-validation.js';
+export * from './kinds/multiplicity/multiplicity-kind.js';
+export * from './kinds/multiplicity/multiplicity-type.js';
+export * from './kinds/primitive/primitive-kind.js';
+export * from './kinds/primitive/primitive-type.js';
+export * from './kinds/top/top-kind.js';
+export * from './kinds/top/top-type.js';
 export * from './kinds/kind.js';
-export * from './kinds/multiplicity-kind.js';
-export * from './kinds/primitive-kind.js';
-export * from './kinds/top-kind.js';
+export * from './kinds/kind-registry.js';
 export * from './utils/dependency-injection.js';
 export * from './utils/test-utils.js';
 export * from './utils/utils.js';
