+++ packages/typir/src/kinds/class/top-class-type.ts
@@ -1,0 +1,76 @@
+/******************************************************************************
+ * Copyright 2024 TypeFox GmbH
+ * This program and the accompanying materials are made available under the
+ * terms of the MIT License, which is available in the project root.
+ ******************************************************************************/
+
+import { TypeEqualityProblem } from '../../features/equality.js';
+import { SubTypeProblem } from '../../features/subtype.js';
+import { isType, Type } from '../../graph/type-node.js';
+import { TypirProblem } from '../../utils/utils-definitions.js';
+import { createKindConflict } from '../../utils/utils-type-comparison.js';
+import { TopClassKind, isTopClassKind } from './top-class-kind.js';
+import { isClassType } from './class-type.js';
+
+export class TopClassType extends Type {
+    override readonly kind: TopClassKind;
+
+    constructor(kind: TopClassKind, identifier: string) {
+        super(identifier);
+        this.kind = kind;
+    }
+
+    override getName(): string {
+        return this.getIdentifier();
+    }
+
+    override getUserRepresentation(): string {
+        return this.getIdentifier();
+    }
+
+    override analyzeTypeEqualityProblems(otherType: Type): TypirProblem[] {
+        if (isTopClassType(otherType)) {
+            return [];
+        } else {
+            return [<TypeEqualityProblem>{
+                $problem: TypeEqualityProblem,
+                type1: this,
+                type2: otherType,
+                subProblems: [createKindConflict(otherType, this)],
+            }];
+        }
+    }
+
+    override analyzeIsSubTypeOf(superType: Type): TypirProblem[] {
+        if (isTopClassType(superType)) {
+            // special case by definition: TopClassType is sub-type of TopClassType
+            return [];
+        } else {
+            return [<SubTypeProblem>{
+                $problem: SubTypeProblem,
+                superType,
+                subType: this,
+                subProblems: [createKindConflict(superType, this)],
+            }];
+        }
+    }
+
+    override analyzeIsSuperTypeOf(subType: Type): TypirProblem[] {
+        // an TopClassType is the super type of all ClassTypes!
+        if (isClassType(subType)) {
+            return [];
+        } else {
+            return [<SubTypeProblem>{
+                $problem: SubTypeProblem,
+                superType: this,
+                subType,
+                subProblems: [createKindConflict(this, subType)],
+            }];
+        }
+    }
+
+}
+
+export function isTopClassType(type: unknown): type is TopClassType {
+    return isType(type) && isTopClassKind(type.kind);
+}
