+++ packages/typir/src/kinds/top/top-kind.ts
@@ -4,72 +4,11 @@
  * terms of the MIT License, which is available in the project root.
  ******************************************************************************/
 
-import { TypeEqualityProblem } from '../features/equality.js';
-import { InferenceRuleNotApplicable } from '../features/inference.js';
-import { SubTypeProblem } from '../features/subtype.js';
-import { isType, Type } from '../graph/type-node.js';
-import { TypirServices } from '../typir.js';
-import { TypirProblem } from '../utils/utils-definitions.js';
-import { createKindConflict } from '../utils/utils-type-comparison.js';
-import { assertTrue, toArray } from '../utils/utils.js';
-import { isKind, Kind } from './kind.js';
-
-export class TopType extends Type {
-    override readonly kind: TopKind;
-
-    constructor(kind: TopKind, identifier: string) {
-        super(identifier);
-        this.kind = kind;
-        this.defineTheInitializationProcessOfThisType({}); // no preconditions
-    }
-
-    override getName(): string {
-        return this.getIdentifier();
-    }
-
-    override getUserRepresentation(): string {
-        return this.getIdentifier();
-    }
-
-    override analyzeTypeEqualityProblems(otherType: Type): TypirProblem[] {
-        if (isTopType(otherType)) {
-            return [];
-        } else {
-            return [<TypeEqualityProblem>{
-                $problem: TypeEqualityProblem,
-                type1: this,
-                type2: otherType,
-                subProblems: [createKindConflict(otherType, this)],
-            }];
-        }
-    }
-
-    override analyzeIsSubTypeOf(superType: Type): TypirProblem[] {
-        if (isTopType(superType)) {
-            // special case by definition: TopType is sub-type of TopType
-            return [];
-        } else {
-            return [<SubTypeProblem>{
-                $problem: SubTypeProblem,
-                superType,
-                subType: this,
-                subProblems: [createKindConflict(superType, this)],
-            }];
-        }
-    }
-
-    override analyzeIsSuperTypeOf(_subType: Type): TypirProblem[] {
-        // a TopType is the super type of all types!
-        return [];
-    }
-
-}
-
-export function isTopType(type: unknown): type is TopType {
-    return isType(type) && isTopKind(type.kind);
-}
-
-
+import { InferenceRuleNotApplicable } from '../../features/inference.js';
+import { TypirServices } from '../../typir.js';
+import { assertTrue, toArray } from '../../utils/utils.js';
+import { isKind, Kind } from '../kind.js';
+import { TopType } from './top-type.js';
 
 export interface TopTypeDetails {
     /** In case of multiple inference rules, later rules are not evaluated anymore, if an earler rule already matched. */
